+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ##### Enhancements
 
+* None.
+
+##### Bug Fixes
+
+* None.
+
+
+## 2.1.1 (2019-10-04)
+
+##### Enhancements
+
 * Added Carthage support.
 [Ryan Gant](https://github.com/ganttastic)
 [#29](https://github.com/BottleRocketStudios/iOS-Scotty/pull/29)